Restructure PreferencesView (PROB2UI-151)
All preference values are now always editable, they no longer need to
be double-clicked to bring them into an "editing" state. Changes to the
preference values are also detected immediately, which prevents changes
getting lost if the user doesn't press Enter in a text field.

File paths are currently only editable as a text field, and not with a
file browser like before. This should be fixed later, perhaps by adding
a "Browse" button next to the text field.