Cache the filtered machine structure in StatesView
Previously if there was a filter, the machine structure would always
be filtered "on the fly" when it was displayed in the tree view.
This caused some issues, such as tree node expandedness resetting
itself when changing trace while there was a filter. This is now fixed,
and performance should be a bit better.