Make GroovyInterpreter eval exception log messages more detailed
There are now separate try blocks for the user code evaluation and the
toString call on the result object. On each try block, there are
separate catch clauses for regular (script/runtime) exceptions, Errors,
and unexpected Throwables.
Currently this only changes the log message, but this could be changed
to display different messages to the user as well.