Refactor MachinesItem refresh code
There is now a single refresh method that refreshes the running icon,
name and last used preference. Refreshing happens automatically when
possible (when the current machine changes). In some cases refresh
still needs to be called manually, for example when the last used
preference changes (because it's not a property).