Extract version info loading from AboutBox into a VersionInfo class
This way other classes can get the version info without having to load
the about box.
The old getCliVersion method also had a race condition: if it was
called before the AboutBox background thread has loaded the version,
it would return null. (This was very unlikely to happen in practice
though.)