Use a single shared empty state spaces for all places that need one
Previously a new empty state space was created every time one was
needed. This was not necessary, because the empty state spaces are
never modified, except by the global preferences view (and that should
affect all empty state space users). By sharing a single empty state
space, fewer probcli instances need to be started, which speeds up the
startup time of the UI.