Refactor and improve ExecuteByPredicateStage
The view for "building" the predicate from a list of name/value pairs
plus an additional predicate is now in a separate PredicateBuilderView
class. This could also be used in other places that a predicate, such
as the "find valid state" dialog.

It is now possible to leave a value or the additional predicate empty,
which causes it to be omitted from the final predicate.