Refactor plugin system to use Path instead of File
This allows removing the warning about a missing inactive.txt when the
ProB 2 UI is run for the first time. If an actual error occurs when
creating the inactive.txt, it is now thrown as an exception.
(Previously the method responsible only returned null, so it was not
possible to tell what kind of error occurred.)