Change how classes to inject in FXML files are detected
Previously, all classes that should be injected using Guice when used
in an FXML file had to be listed in ProB2Module. This has a few
disadvantages (it requires a long list that has to be manually updated,
and singletons are created earlier than necessary because of the
explicit binding). These classes are now annotated with @FXMLInjected,
which is recognized by our custom FXML loader. This removes the need
for the explicit bindings in ProB2Module.