Use a listener system for loading/saving config data
Classes that need to load/save config data now inject the Config class
instead of the other way around. This way, loading the config does not
automatically load all classes that use the config, even if they are
not shown yet (for example the Groovy console). It also makes the code
structure cleaner, because the config loading/saving code is now in the
class that it belongs to, instead of everything being grouped into the
Config class.