Change logging initialization a little
The logback config property is now set in Application.init rather than
in a static initializer block, so that it only takes effect when the
application is actually launched (and not when the class is only
loaded). As a side effect this means that the logger has to be
converted to an instance field instead of a static final field.