Separate visualization preferences from global preferences
Previously the visualization preference views shared their internal
state with the global preference view, which meant that preference
changes would not take effect until the machine was manually reloaded.

The ProBPreferences and PreferencesView classes were also previosuly
marked as singletons, which caused issues with PreferencesDialog
(which needs to use a PreferencesView that is separate from the one in
the preferences window).