Fix some issues with the stage minWidth/minHeight bindings
When the scene's width/height is 0, no minimum is set. This fixes a
StackOverflowError with stages that have an explicit width/height set
in FXML (because in that case the stage's size is initialized before
the scene's).

The stage width/height property is now also listed as a dependency of
the region minWidth/minHeight binding, so that the stage
minWidth/minHeight is updated whenever the stage is resized, even if
the value of the region's minWidth/minHeight property didn't change
(but the calculated minWidth/minHeight might have).