Change machine loading to use ModelFactory instead of Api methods
The ModelFactory classes provide a generic interface for loading models
of different types, so we don't have to write our own interface for
that purpose anymore. They are also more flexible (for example they
allow loading a model directly from a string) and let us show parsing
and loading as separate steps in the status bar.