Refactor view detaching to fix NPEs on Mac
The old detaching code sometimes assumed that the first child of a
detached stage's StackPane is always the detached view. On Mac, this is
not the case (the Mac menu bar takes up that position).

This refactoring adds a new Stage subclass for detached view stages,
which stores the detached view as a field. This allows reliably getting
the detached view from the stage, and also makes the code a little more
organized.