Remove separate classes job from .gitlab-ci.yml
The intention was to have the classes job only build the classes, and
the test job test the already built classes. However, Gradle's
incremental build information is not transferred between build stages,
so the test job has to rerun the entire build anyway. This leads to a
longer build time with no real benefit.