Remove redundant CurrentProject.update calls in checking views
They were apparently meant to mark the project as unsaved when a
checking item is added or removed. This no longer needs to be done
manually - the Machine class automatically marks itself as changed when
a checking item is added or removed, which causes the project to be
marked as unsaved.