Move CurrentProject machine/preference change listeners
Previously the listeners were added in the update method, now this is
done in the CurrentProject change listener. In practice it makes no
difference, because all changes to CurrentProject are done through the
update method, but this ensures that the listeners are always added,
no matter which method is used to make the change.