Remove no longer needed refresh call in StatesView.updateRoot
This was previously necessary because the displayed value in cells
could change without the tree item value changing. (The tree item
previously contained a formula, and the cells displayed the formulas
evaluated value. When the current trace changed, the evaluated values
would change, but the formulas didn't, so the update had to be
triggered manually.)

Now the tree item values are ExpandedFormula objects, which are
replaced with every trace change, so the tree view notices
automatically when it needs to update itself.