Disallow using the prob2-ui build as a subproject
Now when the prob2-ui project is used as a subproject, a message is
displayed that explains how to use it as a composite build instead.

The reason for this change is that subprojects are not the right tool
in this case. Subprojects are meant for structuring separate parts of a
single codebase. To pull in another unrelated codebase as a dependency,
a composite build should be used. Composite builds behave better when
the library in question also exists as a Maven dependency (a composite
build automatically takes priority over a Maven dependency with the
same name, without having to adjust the dependency statement).
Subprojects also cannot use some features that root projects need to
use, such as the Gradle wrapper configuration block.