Fix file path/URL handling in help system
The previous code sometimes converted URIs/URLs to File incorrectly,
which could result in URL strings with multiple file: protocols and
double-escaped parts. This has now been fixed and the corresponding
workarounds removed.

The type of help file/directory URLs is now String (it was previously
File, which was not always correct), this should probably be changed to
URI in the future.