Change help path type from URL strings back to File
Other parts of the code rely on the help files being actual files, so
it makes more sense to use File consistently everywhere.

In the future it might still be a good idea to change the type to URI,
once we no longer depend on the help files being actual files.