Change test case generation trace saving to use DirectoryChooser
This also means that our code no longer needs to handle creating of
directories, because DirectoryChooser only lets users select
directories that already exist.

Existing directories are no longer deleted and recreated, as this could
delete non-test-case files by accident. Instead, a warning is shown if
the selected directory already contains test case traces, asking the
user if they should be overwritten. Non-test-case files in the
directory are ignored (they are not deleted and do not cause a
warning).