Add initial implementation of future-proof JSON reading/writing
The new JsonManager class is used to read/write JSON data with attached
metadata. This metadata contains information such as the type of data,
a version number for the data format, and some other informational
metadata. The format version number is used to detect and
convert/update old data formats, and to prevent loading of newer,
incompatible formats. The type information is used to prevent loading
data that has a different structure than expected.

All JSON files written by the UI now contain this metadata block. The
structure of the metadata is identical for all file types. This also
replaces the old metadata format used in trace and LTL formula files,
which was already written by previous UI versions, but was never
actually read until now. If old metadata is present in a file, it is
read and internally converted to the new format. Newly written files
only use the new metadata format.