Move trace and LTL file validity checks into convertOldData
The check is also only performed if the file type in the metadata is
null (i. e. for old files without metadata). When metadata is present,
the manual validity check should no longer be necessary, as the file
type and format version in the metadata are used to detect incompatible
data.