Change LTL formula file extension from .ltl to .prob2ltl
This matches the naming scheme of all of our other file extensions, and
reduces confusion with plain text .ltl files, as found in the ProB
examples.
Files with the old .ltl extension can still be loaded (this obviously
only works if the files are actually ProB 2 LTL files and not plain
text ones), but new files can only be saved with the new .prob2ltl
extension.