Change StateSview to an initial implementation of lazy evaluation
Now only the currently visible parts of the state view tree are
evaluated and have their subformulas calculated.

This is an early WIP implementation - it does not implement searching,
and does not keep the expanded state of tree items during animation.
It also doesn't implement any kind of caching or reuse of already known
values like the previous implementation did, but this might not be
necessary anymore, depending on how well the performance is improved by
lazy evaluation.