Replace StateItem.current/.previous with individual field accessors
The label and subformulas never change across different states in the
same state space, so there's no need to differentiate between current
and previous in that case.
This simplifies most code that accesses StateItem, and also allows
future optimizations, like caching the tree structure across states so
that it can be accessed without having to re-evaluate all formulas in
the tree.