Separate state error displaying from FullValueStage
State errors are quite different from regular formulas. There's no
current/previous value (the error only applies to a single state), so
the tabs and diff feature are not relevant. The error description is
human-readable text, so the formatting/pretty-print feature is not
useful. It makes more sense to add a separate stage class for state
errors instead of trying to make the full formula value stage work for
state errors too.