Change help system to no longer load help pages from the file system
Previously the structure of the help page tree was directly based on
the file structure of the help page HTML files. This required
converting Java resource URLs (from getResource) to File/Path objects
to list the contents of directories (because Java's resource system
doesn't support listing directories). However, Java resources don't
always correspond to real files on the file system, for example when
they are stored in a jar file. As a workaround in this case, the help
system extracted the help files from the jar onto the file system
(into the prob2ui directory under the ProB home), so that the help
directories could be listed programmatically.

This approach has a few disadvantages:

* It requires two different code paths for when running from source and
  from a jar, which is difficult to test and makes it easy to
  accidentally write code that only works in one of the two cases.
* It makes some assumptions that are true in practice, but not actually
  guaranteed by Java - for example that resource URLs always point
  either into a jar file or onto the real file system.
* When running from a jar, the help files have to be re-extracted onto
  the file system on every run. (This is not a big issue in practice
  though - we have only a few help files and they are quite small.)

This change reimplements the help tree building code in a different way
that avoids these issues. The help tree structure is now created from
resource bundles (.properties files) that list the resource names of
all help pages and describe their tree structure. This way the tree
building code can work entirely using getResource and no longer needs
to list directories using File/Path, which allows the help extraction
code to be removed.

The use of .properties files to list the help pages also has some other
advantages. It allows using Java's native localization support
(resource bundles) to handle help in different languages with fallback
to English, which previously had to be done manually. The new mechanism
also performs the localization and fallback separately for each help
page, which allows translating only some of the help pages while
letting the missing ones fall back to English. (This was previously not
possible, meaning that to add a new help language, all help pages had
to be translated at once.)