Remove Comparable interface from BEditor.Range
The Comparable implementation was not consistent with equals, which is
the Comparable documentation does not recommend. The implementation was
also incomplete (it only compared the start field and not the end and
key fields). The comparison behavior was only used in a single place to
sort a list, so it is easy to replace with a Comparator.