Simplify BEditor executor management
The executor was effectively always running anyway, because the BEditor
always has text in it. The logic for starting/stopping the highlighting
executor was never used except when starting/closing the UI. It's
simpler to always start the executor when BEditor is initialized and
stop it when the UI is closed - this way the BEditor code doesn't have
to handle the case of the executor possibly not existing.