Change BooleanCell back to an icon
The checkbox looked too much like it might be clickable. We now use an
icon again, but a gray check mark instead of a green one, and a blank
space instead of a cross for the false/off state. This hopefully makes
the icon more distinct from the checking status.