Make BEditor error highlighting more reliable
The error items are now stored unfiltered in BEditorView and are
re-filtered when the current editor file changes. This way errors no
longer get "lost" if the machine in question is not currently visible
in the editor. Changing the editor text also no longer clears the saved
error items (they are only hidden), which avoids accidentally clearing
the errors if the file content is loaded after errors are added. Errors
are now only cleared once a machine is (re)loaded.