Optimize HistoryView transition rendering
Each transition's string representation is now created lazily once it
is actually requested, instead of rendering all transition
representations eagerly every time the trace changes.

This significantly improves the history view's performance for long
traces, because most entries are never in view and don't have to be
rendered. However, there is now a small lag when scrolling in the
history view, regardless of the length of the trace.

Fixes PROB2UI-311.