Change StatsView.update to take a StateSpace instead of a Trace
The stats are calculated over the entire state space, not a specific
trace.

This change allows removing some code that gets a trace that wouldn't
otherwise be necessary. This is especially significant in
ModelCheckStats, where the trace was obtained from the checking result.
This is an expensive operation, because the trace is not directly
available and has to be calculated first.