Avoid double calculation of model checking error traces
Previously when a single model checking item was checked, an error was
found, and the UI automatically created a trace to the error state,
that trace was not saved on the item. This meant that when the user
afterwards manually requested to go to the error state, the trace would
have to be recalculated. This is now fixed.

This fixes the remainder of PROB2UI-311.