Allow building a multi-platform shadow jar on Java 11 and newer
This jar has the JavaFX dependencies for all supported operating
systems (Windows, macOS, Linux) added, not just the ones for the
current OS, as is done by the JavaFX plugin. This is not officially
supported or guaranteed to work, so this is done in a separate
multiPlatformShadowJar task instead of the default shadowJar task.