Update StatsView using StateSpace states calculated listener
This replaces the manual update calls in various places that were
previously necessary.

If the StatsView is not updated correctly after a certain operation,
this should be fixed by making the corresponding ProB 2 command
implement IStateSpaceModifier, so that ProB 2 knows that the command
may expand the state space and calls the appropriate listeners.