Replace Modelchecker.resultProperty with states calculated listener
The property was only used by DynamicCommandStage to detect when model
checking has finished, so that the visualisation can be updated (needed
mainly for the state space visualisations, so that states that were
newly explored during model checking are displayed). This is no longer
needed now that we can add listeners for when the state space has
expanded.