Allow other classes to add expressions to DisablePropertyController
This way the parts of the disableProperty expression no longer need to
be listed in a single place, which is more flexible and avoids cyclic
dependency problems (when a class affects the disableProperty but also
wants to use the disableProperty itself).

The implementation is based on a similar mechanism in StatusBar that is
used to show the "Updating views..." message.