Remove redundant hideStats calls in ModelcheckingView
The statsBox is set as not visible in the FXML, so it does not need to
be hidden in initialize. When the machine is reset/changed/unloaded,
hideStats is called by the tvChecks selection listener, which is fired
because the job items list is cleared, which causes the selection in
tvChecks to be cleared.