Simplify trace/counterexample handling in verifications result handlers
Previously the result handler subclasses would call handleFormulaResult
in the abstract superclass to convert the result to a
CheckingResultItem and (in the case of a counterexample result) return
the traces for the result. handleFormulaResult calculated the traces by
calling handleCounterExample, which was an abstract method implemented
in the subclasses. The traces returned from handleCounterExample were
directly returned by handleFormulaResult and not processed further.

This code flow can be simplified by making the subclasses calculate the
traces directly, instead of asking the superclass to do it, which
immediately calls a method in the subclass again.