Replace TestCaseGenerationItem.additionalInformation map with classes
This ensures that only the expected types and values for each test case
generation type are stored.

Although additionalInformation is (de)serialized from/to JSON, this
does *not* require a Project format version bump, because the JSON
format of the new AdditionalInformation subclasses is compatible with
the previous map-based format.