Change LTL pattern parsing to not modify existing pattern item names
When creating or editing a pattern, the pattern is now parsed before
the LTLPatternItem is created. This allows passing the correct name
into the LTLPatternItem constructor, instead of having to first create
the item with a blank name and then replace it with the correct name
once it has been parsed.