Replace automatic minimum stage sizing with manual minimum sizes again
Without native support from JavaFX, it's almost impossible to make
stages automatically respect the minimum size of their content.
Our previous code for this worked some of the time, but would misbehave
or throw exceptions in some situations on some platforms.

Having to set the minimum stage sizes everywhere is not very nice -
you need to manually find out what would be a good minimum size for the
stage, check that it's actually big enough for the content (otherwise
when a user makes the stage small enough, some of the content gets cut
off), and remember to keep it updated whenever the stage's content size
changes. But at least this way it works without errors or other
problems on all platforms.

Fixes PROB2UI-389.