Replace state errors tab with a stage openable via status bar
The state errors list is not used very often, so it's better to make it
open on demand instead of always taking up a tab in the central view.

This also removes the ability to save a state error's long description
to a text file. This feature was carried over from the old
implementation as part of the state view and is probably almost never
used. (If necessary, the text can be copied from the description text
area and manually saved into a file.)