Store config and plugins in location independent of ProB home
This way the UI's user data no longer gets lost when the ProB home
changes, which usually happens when a new version of ProB 2 is
released.

Data from the old location is automatically copied to the new location,
so no user data should get lost in this change.

Fixes PROB2UI-243.