Set proper owners for all dialogs/alerts where possible
Previously we didn't set owners at all in most cases. This caused
problems when using the UI on a system with multiple displays -
ownerless dialogs are always displayed on the primary display, even if
the UI window is on a different display. Dialogs are now owned by the
main stage by default, which gives a better default behavior. Where
possible, dialog owners are set to the specific window that caused them
to be created, which helps when the main stage is on a different
display than the window that created the dialog.