Replace property and listeners in TabPersistenceHandler
There's no need to store the current tab ID in a property - it can
always be easily retrieved from the tab pane, and when the current tab
ID is changed, the tab in question can be selected directly. This way
it's no longer necessary to keep the property and the tab selection
in sync manually.