Simplify checking all operations in symbolic invariant checking
Previously there was a separate symbolic checking type
"Check all operations", but this was not a real item type. When the
user selected the type "Check all operations" in the symbolic checking
choice view, it would instead add multiple separate invariant checking
items, one for each operation in the machine.

This behavior was counterintuitive for the user and complex to
implement in the code, because no other checking type adds multiple
items or items of a different type than the one selected by the user.
The behavior was also strange when changing an existing item to the
"Check all operations" type - this only added new items and didn't
replace/remove the existing item.

This feature is now implemented as part of the normal "Invariant"
symbolic checking type. When adding an item of this type, the operation
choice box now contains a new choice for checking all operations.
Selecting this choice creates a single "Invariant" item with an empty
operation name to indicate that all operations should be checked.