Avoid communicating with CLI in HistoryItem.toPrettyString
If the CLI is executing a command sent by another thread,
toPrettyString blocks until that command finishes executing. Because
toPrettyString is called on the JavaFX application thread, this can
cause the UI to hang, even if the slow command that is blocking the CLI
isn't running on the JavaFX application thread.