Change Guice injector stage to DEVELOPMENT
Stage PRODUCTION tries to load singletons as early as possible, which
makes sense in some use cases (e. g. a server), but in our case
actually makes the UI startup slower. For example, the installation of
probcli binaries previously happened when the injector was created,
which delayed the loading of the UI. Now this happens only once the
empty state space is created, which happens on a background thread
after the main window has been loaded.