Remove preloader and use primary stage as loading window instead
The preloader didn't always work reliably when not running from source
(e. g. when using the jar or the native installers). We only used the
preloader as a workaround to make the loading window display reliably,
because our loading code blocked the JavaFX application thread for a
long time. Now that most of the loading code runs on a background
thread, this is no longer an issue, and we can use the regular primary
stage as the loading window instead.