Replace DynamicCommandStage.reset with more specific methods
interrupt is now used instead of reset when clearing an existing
visualization, so that a visualization that might still be loading in
the background won't be displayed later. There are now also separate
methods for clearing the loading status and visualization content, so
that the loading status can be cleared safely without also removing the
visualization content by accident.