Always enable dot and table visualization menu items
The views can't be used when no machine is loaded, but they don't cause
any errors either, and it's nice to be able to open the views before
loading a machine. It's already possible to get into this state anyway,
by loading a machine, opening the view, then opening another project
(which unloads the machine).