Declare proper menu categories for Linux
On Linux, the application menu isn't a simple folder structure like the
Windows start menu. Instead, every application declares a list of
categories that it belongs to, and the desktop environment organizes
the application menu based on this information (usually grouping
applications by topic). The jpackage option --linux-menu-group
corresponds directly to the list of categories in the generated
.desktop file.

It's generally not possible to create new categories/groups with
arbitrary names. For example, on KDE 5, our previous group setting
(--linux-menu-group ProB2-UI) made the UI show up in a fallback group
called "Lost & Found", not in its own group called "ProB2-UI". Now that
we declare a proper list of standard categories, the UI should be
grouped correctly in the application menu.