Simplify current machine display in title
The machine name in the project is no longer shown in the title,
because this name usually isn't very interesting (it's almost always
the same as the file name anyway, just without the extension).

Also, the machine file name is now always shown, even if the machine is
not fully loaded yet, by getting the file name from the project instead
of from the loaded model.