Re-add most details to log file (but not console log)
The log file is useful for debugging problems that happened before and
perhaps cannot be reproduced easily, so it's better to include more
details here than in the console log. When viewing the log file in an
editor, it's relatively easy to filter out unwanted information using
regex search/replace.