Display dot/table view errors in visualization area
This allows displaying errors for visualization commands that take no
arguments, makes it more obvious that the visualization failed
(compared to just leaving the visualization area blank), and saves a
bit of space in the window. The error display can't conflict with the
actual visualization, because errors are only displayed if the
visualization failed.