Simplify handling of trailing space after console prompts
This space previously had to be entered as part of the console prompt,
which was quite confusing in properties files, because trailing spaces
are difficult to see in most editors. Now the space is added by the
Console code after every prompt and no longer needs to be part of the
prompt string and the properties files. (There should be no need for
prompts *without* a space afterwards.)