Remove owners from non-modal stages
At least on macOS, stages owned by another window are always forced in
front of their owner, even if not set as modal, which is generally not
the behavior we want.

This unfortunately means that these stages will appear on the primary
screen by default again, even if the main UI window is on a different
screen.