Display error to user if an uncaught exception is thrown during startup
Previously, if an error happened before the main UI is fully loaded,
the UI would be stuck forever on the loading window. Also, if the
exception happened before the uncaught exception handler was set up,
the error wouldn't be displayed to the user at all.

This change solves both of these problems: exceptions before the
uncaught exception handler are caught and displayed manually, and this
error display replaces the loading window, so that the UI exits when
the user closes the error window.