Replace CodeMirror LTL editors with RichTextFX CodeArea
The CodeMirror-based editors in WebViews had issues where they wouldn't
load for some users. As a workaround, replace them with a CodeArea from
RichTextFX, as we've already done for the B editor.

The new editors currently don't support syntax highlighting and only do
limited error highlighting. In the future we should try to generalize
our highlighting implementation in BEditor so that it can also be used
in the LTL editors.