Remove description field from non-LTL checking items
For other checking item types the description was not provided by the
user, but derived from another field in the checking item. In these
cases, the field in question is now used directly instead, and there is
no longer a redundant description field in the class and the project
file.