Fix rare "not on UI thread" exception when loading NSMenuFX
The previous code had a race condition - there's no guarantee when a
Platform.runLater callback will actually be executed, so it was
possible (but very unlikely) for other non-UI-thread code to inject a
MenuToolkit object before the callback preloaded the MenuToolkit on the
UI thread. As a workaround, move injector creation back onto the UI
thread and load the MenuToolkit before starting the background loading
thread. Creating the injector takes a second or so on my 2012 MBP,
so this isn't a huge responsiveness problem.