Remove name field from all checking items except LTLPatternItem
For all other checking item types the name was either empty or derived
from other fields, so there's no need to store it as an actual field in
AbstractCheckableItem and in the project file. Even in LTLPatternItem
the name field is derived from the code field - we keep it in the
project file only to avoid calling the LTL pattern parser every time a
project/machine is loaded.