Move AbstractCheckableItem.code field into subclasses
The contents of the code field have a completely different format for
each subclass, so a generic interface isn't very useful here (and it
was actually never used generically, only when the concrete class of
checking item was already known).

This change also allows more flexibility for changing the format of the
code field in the subclasses. For example, some symbolic
checking/animation types take an argument that's better represented as
a list or enum value instead of a string.

Test case generation items already didn't use the code field (it was
always set to an empty string), so in that case the field is removed
completely.