Disallow setting VisBVisualisation fields to null
Previously a VisBVisualisation object with all fields set to null was
used as a placeholder when no visualisation was loaded. Now the entire
VisBVisualisation field is set to null instead, so that a non-null
VisBVisualisation never has null fields. This allows simplifying some
of the logic for checking if a visualisation is loaded.