Restructure VisB SVG/HTML loading to avoid a race condition
Previously the listener for adding click events could be called when
visBVisualisation was still null. I'm not sure how exactly that could
happen, because a page load should only be started after
visBVisualistion is set...

This change moves the addClickEvent calls into a runWhenLoaded block
inside VisBStage's visBController listener, to ensure that the
visualisation object is available when the code runs. To make this
work, the SVG/HTML loading also has to be moved into the listener -
otherwise the page would only start loading after the runWhenLoaded
call, so the code in the runWhenLoaded block would execute in the
context of the previous visualisation. This is a good change anyway,
because it removes another reverse dependency of VisBController on
VisBStage.