Make TestCaseGenerationItem.type not a real field
The type can be derived from the class, so it doesn't need to be stored
as a field and can instead be returned from a method implemented in the
subclasses. This allows us to use the standard Jackson type handling
mechanism, where the JSON type property is used only by Jackson and not
shared with a real Java field. This also decouples the JSON type names
from the TestCaseGenerationType enum constant names.