Remove old ValidationTask class and editing view
All validation task info is now stored in the checking items. The VT
table in the VO manager is still there for now, as an easy overview of
all VTs in a project (every checking item with a non-null ID is
considered a VT).

There is currently no way to edit the VT IDs of checking items. This
still needs to be implemented.