Always update checked status of all VOs
Previously each VO's checked status was only bound when it was checked
for the first time. With this change, it's more obvious when two VOs use
the same task(s) and so don't have to be re-checked.

Also improves handling of VOs with an invalid validation expression -
they now show a parse error status instead of throwing an internal
exception when checking them.