Always check newly added symbolic/test case/simulation items
The second button for adding the item without checking it is removed,
matching how the model checking and LTL formula dialogs already work.
This makes the UI a bit simpler to understand and simplifies some
internal code. In practice, the difference doesn't matter much - usually
the check is quite fast, and if not, the user can cancel it.