Move most code from SymbolicFormulaHandler to concrete classes
This bit of code is now duplicated in the symbolic formula and animation
classes, which isn't nice, but it's a bit less tangled this way.

This should be refactored again at some point, but in a way that covers
more than just the symbolic classes - e. g. LTLView also contains
similar code for adding/changing formulas.