Fully separate symbolic checking and animation choosing stages
This (again) duplicates a bit of code, but allows more flexibility in
the implementation. Some parts were already very specific to one of the
two stages and can be removed in the other one. The remaining parts can
be better customized in the future to fit the specific checking types
better.