Display checking cancel buttons only while a check is running
Now all check buttons behave like the random animation button in the
operations view: by default, only the button to start the check is
shown, and while a check is running, it's replaced with the button to
cancel the running check. This simplifies the toolbar without losing
functionality, because the two buttons are never enabled at once.

The logic for this is now in a separate helper class, which simplifies
the code of the individual checking views a bit.