Show detaching context menu only for pane title, not content
The previous workaround (specific to the visualization pane) had the
same effect, but unfortunately also disabled the context menus from the
pane content, e. g. the right-click options in the animation function
visualization.

With this solution, all context menus in the pane content work
correctly, but it has the disadvantage that the detaching menu only
works on the pane title bar and not on other blank spaces in the pane.